Moved TextControl/Loader to plugins
