+++ frictionless/__init__.py
@@ -32,7 +32,6 @@ from . import controls
 from . import dialects
 from . import errors
 from . import exceptions
-from . import loaders
 from . import parsers
 from . import steps
 from . import types
