--- frictionless/loaders/__init__.py
@@ -1,1 +1,0 @@
-from .text import TextLoader
