+++ tests/plugins/test_text.py
